remove unnecessary unboxing
Remove explicit unwrapping of wrapped primitive values. Unboxing is unnecessary
under Java 5 and newer.

+++ src/main/java/net/blerf/ftl/model/ShipLayout.java
@@ -66,7 +66,7 @@ public class ShipLayout {
 	public int getRoomCount() {
 		try {
 			Integer lastKey = roomMap.lastKey();
-			return lastKey.intValue()+1;
+			return lastKey +1;
 		}
 		catch ( NoSuchElementException e ) {
 			return 0;
